fleetctl: destroy a unit before replacing a unit
In order to replace a unit with a new one, fleetctl should delete an
existing unit before creating a new one. Otherwise the old unit would
remain with a state either "loaded" or "launched". Then a command with a
replace option could be hanging, because checkUnitState() waits for an
"inactive" state indefinitely.