functional: introduce new helpers for testing the replace option
Add new helpers copyFile(), genNewFleetService() & destroyUnitRetrying()
to prepare the new functional tests for the replace options.

copyFile() is a helper to copy one file to another.
genNewFleetService() is a helper to replace a string with a new one.
It's necessary for the next functional tests.
destroyUnitRetrying() runs "fleetctl <cmd> --replace" repeatedly to ensure
the unit is actually removed.