registry: remove extra loops when fetching Jobs list from etcd
Improve Units() call when we fetch and process Jobs and Units from
etcd. Remove extra unnecessary loops.

We also change the code logic a little bit, since it was always storing
the last matched name with the new Unit, but since Job keys do not expect to
have two units with the same name, this should be ok.