fleetctl: add checkUnitCreation() and isLocalUnitDifferent()
Add checkUnitCreation() to check if the unit should be created or not.
This function handles the new replace logic.

Add isLocalUnitDifferent() since we don't really want to warn if the
Unit do really differ in case "--replace" switch was set. At the same
time factor our unit matching logic. The function handles both cases
when '--replace' is set and not.